split out the SignJWT functions into their own file and command
